Add Protocol\InsertText to deal with snippets on completion
https://microsoft.github.io/language-server-protocol/specification#textDocument_completion

It's possible to provide the key `insertText` to let the client know
what to insert in place of the pattern typed by the user on completion.
For example:
```
new SuggesITF<>
```

Would be resolve to `SuggestionInsertTextFactory` but the client need to
replace `SuggesITF` by it.
`insertText` is used for that purpose, if `insertText` is not provided
`label` is used instead.

Another key allows to specify the type of the text to insert:
`insertTextFormat`
Which can have two values:
1. Plain text
2. Snippet

To deal with that I added a new prorocol value object `InsertText` which
provide both the text to insert and the type.
I also created a factory that can create an `InsertText` from a
`Suggestion`: `SuggestionInsertTextFactory`

The next step is to inject this factory in the `CompletionHandler` and
generate the appropriate `CompletionItem`
I would like to rewwork the `CompletionItem` in order to provide it only
the `InsertText` object instead of the value and type separatly.
Especially because they are far from each other which is impractical.
This will also reduce the knowledge we have about constructing an
`CompletionItem` by letting the object itself defining its values.